No disk/partition files for configuration removed
